workaround Badge/Key Item/HM dupe UI bug
Dumb hack workaround for a minor UI bug that occurs when multiple Locations (Checks) share the same name but have different properties.
The dumb hack workaround in question is to make the names slightly different, by adding an extra space to the end of one of them.
Additional note: this hack doesn't cause issues with auto-tracking when Badge/Key Item/HM rando is disabled, because I think the state of the Location/Check in the tracker UI (on the map) is synced with the state of the Item (under "Items" in the UI).